# docs/swagger.yaml
openapi: 3.0.0
info:
  title: TradeVault API
  description: |
    A comprehensive trading platform API that provides:
    
    ## Features
    - **User Authentication**: Secure login/register with JWT tokens
    - **OTP Verification**: Email-based OTP system for account security
    - **Real-time Trading**: WebSocket-based live stock updates
    - **Stock Management**: Buy/sell stocks with portfolio tracking
    - **Biometric Authentication**: Advanced security with biometric keys
    - **OAuth Integration**: Google and Apple sign-in support
    
    ## Authentication
    - **App Token**: Used for main application endpoints (JWT_SECRET)
    - **Socket Token**: Used for WebSocket connections (SOCKET_TOKEN_SECRET)
    
    ## Rate Limiting
    - OTP requests: 1 per minute per email
    - Login attempts: 3 attempts before 30-minute lockout
    - PIN attempts: 3 attempts before 30-minute lockout
    
  version: '1.0.0'
  contact:
    name: TradeVault Support
    email: support@tradevault.com
    url: https://tradevault.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.tradevault.com
    description: Production server
security:
  - BearerAuth: []
  - SocketAuth: []
paths:
...
  /stocks:
    get:
      tags:
        - Stock Management
      summary: Get all stocks
      description: Retrieves a list of all available stocks
      operationId: getAllStocks
      security:
        - SocketAuth: []
      responses:
        '200':
          description: Stocks retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetStocksResponse'
        '401':
          description: Unauthorized - invalid socket token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /stocks/stock:
    get:
      tags:
        - Stock Management
      summary: Get stock by symbol
      description: Retrieves detailed information about a specific stock
      operationId: getStockBySymbol
      security:
        - SocketAuth: []
      parameters:
        - name: stock
          in: query
          description: Stock symbol (e.g., AAPL, GOOGL)
          required: true
          schema:
            type: string
          example: "AAPL"
      responses:
        '200':
          description: Stock retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetStockResponse'
        '400':
          description: Bad request - missing stock symbol
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Stock not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /stocks/register:
    post:
      tags:
        - Stock Management
      summary: Register new stock
      description: Adds a new stock to the trading platform
      operationId: registerStock
      security:
        - SocketAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterStockRequest'
            example:
              symbol: "TSLA"
              companyName: "Tesla, Inc."
              iconUrl: "https://example.com/tesla-icon.png"
              lastDayTradedPrice: "250.00"
              currentPrice: 255.50
      responses:
        '201':
          description: Stock registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterStockResponse'
        '400':
          description: Bad request - invalid data or stock already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ========================================
  # TRADING ENDPOINTS (Protected)
  # ========================================
  /stocks/buy:
    post:
      tags:
        - Trading
      summary: Buy stock
      description: Purchases a specified quantity of a stock
      operationId: buyStock
      security:
        - SocketAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuyStockRequest'
            example:
              stock_id: "507f1f77bcf86cd799439011"
              quantity: 10
      responses:
        '201':
          description: Stock purchased successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuyStockResponse'
        '400':
          description: Bad request - insufficient balance or invalid data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /stocks/sell:
    post:
      tags:
        - Trading
      summary: Sell stock
      description: Sells a specified quantity of a stock from holdings
      operationId: sellStock
      security:
        - SocketAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SellStockRequest'
            example:
              holdingId: "507f1f77bcf86cd799439012"
              quantity: 5
      responses:
        '200':
          description: Stock sold successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SellStockResponse'
        '400':
          description: Bad request - insufficient quantity or invalid data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /stocks/holding:
    get:
      tags:
        - Trading
      summary: Get user holdings
      description: Retrieves all stock holdings for the authenticated user
      operationId: getAllHoldings
      security:
        - SocketAuth: []
      responses:
        '200':
          description: Holdings retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetHoldingsResponse'
        '401':
          description: Unauthorized - invalid socket token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token for application authentication.
        Include in Authorization header: `Bearer <token>`
    SocketAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token for WebSocket and trading operations.
        Include in Authorization header: `Bearer <token>`
  schemas:
...
    RegisterStockRequest:
      type: object
      required:
        - symbol
        - companyName
        - iconUrl
        - lastDayTradedPrice
        - currentPrice
      properties:
        symbol:
          type: string
          description: Stock symbol/ticker
          example: "TSLA"
        companyName:
          type: string
          description: Company name
          example: "Tesla, Inc."
        iconUrl:
          type: string
          format: uri
          description: URL to company icon
          example: "https://example.com/tesla-icon.png"
        lastDayTradedPrice:
          type: string
          description: Previous day's closing price
          example: "250.00"
        currentPrice:
          type: number
          description: Current stock price
          example: 255.50
    RegisterStockResponse:
      type: object
      properties:
        msg:
          type: string
          description: Success message
          example: "Stock added successfully!"
        data:
          $ref: '#/components/schemas/Stock'
    GetStocksResponse:
      type: object
      properties:
        msg:
          type: string
          description: Success message
          example: "Stocks retrieved successfully!"
        data:
          type: array
          items:
            $ref: '#/components/schemas/Stock'
    GetStockResponse:
      type: object
      properties:
        msg:
          type: string
          description: Success message
          example: "Stock retrieved successfully!"
        data:
          $ref: '#/components/schemas/Stock'

    # ========================================
    # TRADING SCHEMAS
    # ========================================
    BuyStockRequest:
      type: object
      required:
        - stock_id
        - quantity
      properties:
        stock_id:
          type: string
          description: Stock identifier
          example: "507f1f77bcf86cd799439011"
        quantity:
          type: integer
          minimum: 1
          description: Number of shares to buy
          example: 10
    BuyStockResponse:
      type: object
      properties:
        msg:
          type: string
          description: Success message
          example: "Stock bought successfully!"
        data:
          $ref: '#/components/schemas/Holding'
    SellStockRequest:
      type: object
      required:
        - holdingId
        - quantity
      properties:
        holdingId:
          type: string
          description: Holding identifier
          example: "507f1f77bcf86cd799439012"
        quantity:
          type: integer
          minimum: 1
          description: Number of shares to sell
          example: 5
    SellStockResponse:
      type: object
      properties:
        msg:
          type: string
          description: Success message
          example: "Stock sold successfully!"
        data:
          type: object
          properties:
            orderId:
              type: string
              description: Order identifier
              example: "507f1f77bcf86cd799439013"
            sellPrice:
              type: number
              description: Total sale amount
              example: 1277.50
    GetHoldingsResponse:
      type: object
      properties:
        msg:
          type: string
          description: Success message
          example: "Holdings retrieved successfully!"
        data:
          type: array
          items:
            $ref: '#/components/schemas/Holding'
...
    Stock:
      type: object
      properties:
        _id:
          type: string
          description: Unique stock identifier
          example: "507f1f77bcf86cd799439011"
        symbol:
          type: string
          description: Stock symbol/ticker
          example: "TSLA"
        companyName:
          type: string
          description: Company name
          example: "Tesla, Inc."
        iconUrl:
          type: string
          format: uri
          description: URL to company icon
          example: "https://example.com/tesla-icon.png"
        lastDayTradedPrice:
          type: string
          description: Previous day's closing price
          example: "250.00"
        currentPrice:
          type: number
          description: Current stock price
          example: 255.50
        dayTimeSeries:
          type: array
          description: Intraday price data
          items:
            type: object
        tenMinTimeSeries:
          type: array
          description: 10-minute interval price data
          items:
            type: object
    Holding:
      type: object
      properties:
        _id:
          type: string
          description: Unique holding identifier
          example: "507f1f77bcf86cd799439011"
        user:
          type: string
          description: User identifier
          example: "507f1f77bcf86cd799439012"
        stock:
          type: string
          description: Stock identifier
          example: "507f1f77bcf86cd799439013"
        quantity:
          type: integer
          description: Number of shares held
          example: 10
        buyPrice:
          type: number
          description: Purchase price per share
          example: 250.00
        createdAt:
          type: string
          format: date-time
          description: Holding creation timestamp
          example: "2024-01-01T00:00:00.000Z"
...
    ErrorResponse:
      type: object
      properties:
        msg:
          type: string
          description: Error message
          example: "Invalid request data"
...